make all unsigneds into shorts
In keeping with the original spirit of the code. Not that I intend to
merge my changes with the original or anything.

+++ opl2test/timer.h
@@ -1,7 +1,7 @@
 #ifndef TIMER_H_
 #define TIMER_H_
 
-void timer_setup(unsigned frequency);
+void timer_setup(unsigned short frequency);
 void timer_shutdown();
 unsigned long timer_get();
